requirements.txt
dbt-snowflake==1.7.0
dbt-utils==1.1.1
pydantic==2.0.0
python-dotenv==1.0.0
snowflake-connector-python==3.2.0
pyspark==3.5.0
databricks-sdk==0.20.0
apache-airflow==2.8.0
apache-airflow-providers-slack==7.1.0
anthropic==0.25.0
great-expectations==0.18.12
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.0
pylint==3.0.3
sqlfluff==2.2.0
pydantic-settings==2.0.3
